requirements.txt
mediapipe
opencv-python
paho-mqtt
tensorflow[and-cuda]